+++ src/actions/introAction.js
@@ -1,6 +1,6 @@
 import { INTRO } from './types';
 
-export const changeIntro = intro => dispatch => {
+export const changeInt = intro => dispatch => {
   dispatch({
     type: INTRO,
     payload: intro
