+++ src/reducers/introReducer.js
@@ -7,6 +7,7 @@ const initialState = {
 export default function(state = initialState, action){
   switch (action.type){
     case INTRO:
+      console.log("here")
       return {
         ...state,
         intro: action.payload
